requirements.txt
discord.py[voice]>=2.3.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
